feat(slider): can use VoiceOver to interact
VoiceOver couldnt not update the slider because
the slider is not listening for events triggered
by voiceover (gesture). Gesture is triggered only
on native elements like input.
this new implementation uses the input of type
range as the target of the focus and add a new
event listener for gesture.
Other future features could allow for simply
listening for changes to the html input to be
reflected to the visual slider. Nothing should be
changed when it comes to keyboard and mouse
behavior.

Fixes #15037